requirements.txt
chainlit>=1.1.0
openai>=1.30.0
httpx>=0.27.0
chromadb>=0.5.0
beautifulsoup4>=4.12.3
tenacity>=8.4.1
pydantic>=2.7.1
python-dotenv>=1.0.1
urllib3>=2.2.1
markdownify>=0.13.1
